refactor(negotiation): package refactor
