+++ a2a-t-negotiation/src/main/java/net/openan/a2at/sdk/negotiation/package-info.java
@@ -1,0 +1,2 @@
+/** Negotiation support for the a2a-t Java SDK. */
+package net.openan.a2at.sdk.negotiation;
