+++ a2a-t-negotiation/src/test/java/net/openan/a2at/sdk/negotiation/runtime/NegotiationRuntimeStateMachineTest.java
@@ -3,7 +3,7 @@ package net.openan.a2at.sdk.negotiation.runtime;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.Map;
-import net.openan.a2at.sdk.negotiation.runtime.impl.NegotiationRuntime;
+
 import net.openan.a2at.sdk.negotiation.store.impl.InMemoryNegotiationStore;
 import net.openan.a2at.sdk.negotiation.types.exception.NegotiationStateException;
 import net.openan.a2at.sdk.negotiation.handler.ClarificationNegotiation;
